Move includes/formatters to src/Formatters
Also correctly namespace the files including class name.

The old classname is still available but deprecated.
